Avoid glob pattern `**` to support more shell environments and Node.js versions.
Unfortunately, this means all test modules need to be in the repo root directory. Luckily that’s the case, for now. In Node.js v21+ the test runner CLI can accept advanced globs, but it will be a long time until this project can drop support for Node.js v18 and v20.